--- src/interface/README.md
@@ -1,5 +1,0 @@
-# ***Camada de Apresentação***
-
-
-* Responsável por toda a interação visual com o usuário. Contém componentes e páginas da aplicação, estruturando a forma como os dados e funcionalidades são exibidos. Pode conter arquivos como layouts, inputs, filtros e dashboards interativos.
-Tecnologias comuns: Streamlit (com HTML/CSS customizados), componentes de visualização (como Plotly ou matplotlib se necessário).
