--- src/interface/views/README.md
@@ -1,3 +1,0 @@
-# ***Views (telas da aplicação)***
-
-* Subdiretório da interface que organiza as telas principais e seções da aplicação (como visão geral, relatório financeiro, alertas, etc). Ajuda na separação de responsabilidades por contexto visual e facilita a manutenção.
