--- src/interface/views/analises.py
@@ -1,6 +1,0 @@
-
-import streamlit as st
-
-def analises_page():
-    st.title("Análises Inteligentes")
-    st.write("Conteúdo da página de Análises Inteligentes")
