+++ src/interface/views/assets/img/equipe/Eduarda.png
