+++ src/interface/views/assets/img/equipe/eric.png
