+++ src/interface/views/assets/img/equipe/guilherme.png
