+++ src/interface/views/assets/img/home_ilustracao.png
