+++ src/interface/views/assets/img/sobre_ilustracao.png
