--- src/interface/views/dados.py
@@ -1,6 +1,0 @@
-
-import streamlit as st
-
-def dados_page():
-    st.title("Dados")
-    st.write("Conteúdo da página de Dados")
