+++ src/interface/views/static/font/Inter-VariableFont.ttf
